Split core functionality into pdf_engine.py
this change is necessary to prevent the vestigial import of FreeSimpleGUI in the flask app.
